+++ src/features/levels/index.ts
@@ -4,5 +4,5 @@ export { RoundState, type MoveCardAction, RowType } from "./model/RoundState";
 export { StageStatus } from "./model/Stage";
 export { type Stage } from "./model/Stage";
 
-export { RoundControls } from "./ui/RoundControls";
+export { LevelControls } from "./ui/LevelControls";
 export { StageControls } from "./ui/StageControls";
