+++ src/features/levels/ui/LevelControls.ts
@@ -4,11 +4,10 @@ import { Observer, Publisher } from "../../../entities/state";
 
 import { LevelsState } from "../model/LevelsState";
 
-import styles from "./RoundControls.module.css";
+import styles from "./LevelControls.module.css";
 
 // TODO: create generic controls class?
-// TODO: rename to LevelsControls
-export class RoundControls extends Component implements Observer {
+export class LevelControls extends Component implements Observer {
   difficultySelect: Component;
 
   roundSelect: Component;
