+++ src/widgets/game-fields/index.ts
@@ -1,2 +1,2 @@
-export { GameField } from "./ui/sentence-board/GameField";
+export { SentenceBoard } from "./ui/sentence-board/SentenceBoard";
 export { WordsPicker } from "./ui/words-picker/WordsPicker";
