+++ src/widgets/game-fields/ui/sentence-board/SentenceBoard.module.css
